requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django @ file:///tmp/build/80754af9/django_1601563330930/work
django-crispy-forms==1.9.2
djangorestframework==3.12.1
idna==2.10
psycopg2 @ file:///tmp/build/80754af9/psycopg2_1594305108220/work
pytz==2020.1
requests==2.24.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
urllib3==1.25.11
